Folder reorganisation
The folder “tests” is now inside the root, that’s because the tests are
used both for the compiler and the virtual machine. This makes the
“source” folder inside the “compiler” folder useless, hence it has been
removed and its files now are simply inside “compiler”.
I’m not sure whether or not the folder “Game of Life” (which has been
stripped of the underscores) should stay inside the root or the “tests”
folder.